Remove addmm node check from linear pass and combine conv/linear fusion test files
